Update server core to RunUO 2.7

+++ Scripts/Server/Accounting/AccountHandler.cs
@@ -8,7 +8,6 @@ using Server.Commands;
 using Server.Engines.Help;
 using Server.Network;
 using Server.Regions;
-using Server.Mobiles;
 
 namespace Server.Misc
 {
@@ -22,7 +21,8 @@ namespace Server.Misc
 	public class AccountHandler
 	{
 		private static int MaxAccountsPerIP = 10;
-
+		private static bool AutoAccountCreation = true;
+		private static bool RestrictDeletion = false;
 		private static TimeSpan DeleteDelay = TimeSpan.FromDays( 7.0 );
 
 		public static PasswordProtection ProtectPasswords = PasswordProtection.NewCrypt;
@@ -33,11 +33,6 @@ namespace Server.Misc
 		{
 			get{ return m_LockdownLevel; }
 			set{ m_LockdownLevel = value; }
-		}
-
-		public static bool AutoAccountCreation()
-		{
-			return true;
 		}
 
 		private static CityInfo[] StartingCities = new CityInfo[]
@@ -53,6 +48,23 @@ namespace Server.Misc
 				new CityInfo( "Vesper",		"The Ironwood Inn",	1075080, 2771,	976,	0  )
 			};
 
+		/* Old Haven/Magincia Locations
+			new CityInfo( "Britain", "Sweet Dreams Inn", 1496, 1628, 10 );
+			// ..
+			// Trinsic
+			new CityInfo( "Magincia", "The Great Horns Tavern", 3734, 2222, 20 ),
+			// Jhelom
+			// ..
+			new CityInfo( "Haven", "Buckler's Hideaway", 3667, 2625, 0 )
+
+			if ( Core.AOS )
+			{
+				//CityInfo haven = new CityInfo( "Haven", "Uzeraan's Mansion", 3618, 2591, 0 );
+				CityInfo haven = new CityInfo( "Haven", "Uzeraan's Mansion", 3503, 2574, 14 );
+				StartingCities[StartingCities.Length - 1] = haven;
+			}
+		*/
+
 		private static bool PasswordCommandEnabled = true;
 
 		public static void Initialize()
@@ -66,7 +78,7 @@ namespace Server.Misc
 		}
 
 		[Usage( "Password <newPassword> <repeatPassword>" )]
-		[Description( "Changes the password of the commanding players account." )]
+		[Description( "Changes the password of the commanding players account. Requires the same C-class IP address as the account's creator." )]
 		public static void Password_OnCommand( CommandEventArgs e )
 		{
 			Mobile from = e.Mobile;
@@ -109,7 +121,7 @@ namespace Server.Misc
 			bool isSafe = true;
 
 			for ( int i = 0; isSafe && i < pass.Length; ++i )
-				isSafe = ( pass[i] >= 0x20 && pass[i] < 0x80 );
+				isSafe = ( pass[i] >= 0x20 && pass[i] < 0x7F );
 
 			if ( !isSafe )
 			{
@@ -119,8 +131,36 @@ namespace Server.Misc
 
 			try
 			{
-				acct.SetPassword( pass );
-				from.SendMessage( "The password to your account has changed." );
+				IPAddress ipAddress = ns.Address;
+
+				if ( Utility.IPMatchClassC( accessList[0], ipAddress ) )
+				{
+					acct.SetPassword( pass );
+					from.SendMessage( "The password to your account has changed." );
+				}
+				else
+				{
+					PageEntry entry = PageQueue.GetEntry( from );
+
+					if ( entry != null )
+					{
+						if ( entry.Message.StartsWith( "[Automated: Change Password]" ) )
+							from.SendMessage( "You already have a password change request in the help system queue." );
+						else
+							from.SendMessage( "Your IP address does not match that which created this account." );
+					}
+					else if ( PageQueue.CheckAllowedToPage( from ) )
+					{
+						from.SendMessage( "Your IP address does not match that which created this account.  A page has been entered into the help system on your behalf." );
+
+						from.SendLocalizedMessage( 501234, "", 0x35 ); /* The next available Counselor/Game Master will respond as soon as possible.
+																	    * Please check your Journal for messages every few minutes.
+																	    */
+
+						PageQueue.Enqueue( new PageEntry( from, String.Format( "[Automated: Change Password]<br>Desired password: {0}<br>Current IP address: {1}<br>Account IP address: {2}", pass, ipAddress, accessList[0] ), PageType.Account ) );
+					}
+
+				}
 			}
 			catch
 			{
@@ -157,7 +197,7 @@ namespace Server.Misc
 					state.Send( new DeleteResult( DeleteResultType.CharBeingPlayed ) );
 					state.Send( new CharacterListUpdate( acct ) );
 				}
-				else if ( DateTime.UtcNow < (m.CreationTime + DeleteDelay) )
+				else if ( RestrictDeletion && DateTime.UtcNow < (m.CreationTime + DeleteDelay) )
 				{
 					state.Send( new DeleteResult( DeleteResultType.CharTooYoung ) );
 					state.Send( new CharacterListUpdate( acct ) );
@@ -214,18 +254,32 @@ namespace Server.Misc
 			}
 		}	
 
+		private static readonly char[] m_ForbiddenChars = new char[]
+		{
+			'<', '>', ':', '"', '/', '\\', '|', '?', '*'
+		};
+
+		private static bool IsForbiddenChar( char c )
+		{
+			for ( int i = 0; i < m_ForbiddenChars.Length; ++i )
+				if ( c == m_ForbiddenChars[i] )
+					return true;
+
+			return false;
+		}
+
 		private static Account CreateAccount( NetState state, string un, string pw )
 		{
 			if ( un.Length == 0 || pw.Length == 0 )
 				return null;
 
-			bool isSafe = true;
+			bool isSafe = !( un.StartsWith( " " ) || un.EndsWith( " " ) || un.EndsWith( "." ) );
 
 			for ( int i = 0; isSafe && i < un.Length; ++i )
-				isSafe = ( un[i] >= 0x20 && un[i] < 0x80 );
+				isSafe = ( un[i] >= 0x20 && un[i] < 0x7F && !IsForbiddenChar( un[i] ) );
 
 			for ( int i = 0; isSafe && i < pw.Length; ++i )
-				isSafe = ( pw[i] >= 0x20 && pw[i] < 0x80 );
+				isSafe = ( pw[i] >= 0x20 && pw[i] < 0x7F );
 
 			if ( !isSafe )
 				return null;
@@ -266,7 +320,7 @@ namespace Server.Misc
 
 			if ( acct == null )
 			{
-				if ( AutoAccountCreation() && un.Trim().Length > 0 )	//To prevent someone from making an account of just '' or a bunch of meaningless spaces 
+				if ( AutoAccountCreation && un.Trim().Length > 0 ) // To prevent someone from making an account of just '' or a bunch of meaningless spaces
 				{
 					e.State.Account = acct = CreateAccount( e.State, un, pw );
 					e.Accepted = acct == null ? false : acct.CheckAccess( e.State );
@@ -358,6 +412,25 @@ namespace Server.Misc
 
 			if ( !e.Accepted )
 				AccountAttackLimiter.RegisterInvalidAccess( e.State );
+		}
+
+		public static bool CheckAccount( Mobile mobCheck, Mobile accCheck )
+		{
+			if ( accCheck != null )
+			{
+				Account a = accCheck.Account as Account;
+
+				if ( a != null )
+				{
+					for ( int i = 0; i < a.Length; ++i )
+					{
+						if ( a[i] == mobCheck )
+							return true;
+					}
+				}
+			}
+
+			return false;
 		}
 	}
 }
